[FIX] delivery: default invoicing policy
Since commit https://github.com/odoo/odoo/commit/df07870a149d54abccca2f00b361ad52c011c1b2
the default invoicing policy of a product is set to 'delivery'. This is
not ok for the delivery products: a delivery product is never delivered.

This is confusing to end users, because they do not realize that a
delivery line is not included in an invoice. This is especially true if
the delivery is free.

This commit sets the default policy to 'order' for master data and for
newly created products from the shipping form view.

opw-2387437

closes odoo/odoo#62357

X-original-commit: 4c2768846809f6687d03eb0d7f501e57c2c93a90
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>