[FIX] uom: move product_uom_dozen back to data
In commit fea27fcc5bd the Dozens unit of measure was moved to demo data.
Unfortunately, this data is used in enterprise module product_unspc in a
post init hook.

As a consequence, when installing this module without demo data, the
initialization crashes.

With this commit, the Dozens uom is moved back to regular data.

closes odoo/odoo#63120

X-original-commit: 6efb7306551fa3f176c8101b6effa095bd365800
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>